+++ apps/web/package.json
@@ -17,6 +17,7 @@
 		"@emotion/server": "^11.10.0",
 		"@juggle/resize-observer": "3.3.1",
 		"@kittr/logger": "workspace:*",
+		"@kittr/firebase": "workspace:*",
 		"@kittr/prisma": "workspace:*",
 		"@kittr/ui": "workspace:*",
 		"@kittr/analytics": "workspace:*",
@@ -34,11 +35,7 @@
 		"@trpc/react": "^9.27.4",
 		"@trpc/react-query": "^10.9.0",
 		"@trpc/server": "^10.9.0",
-		"browser-image-compression": "2.0.0",
 		"date-fns": "2.28.0",
-		"firebase": "^9.16.0",
-		"firebase-admin": "^11.5.0",
-		"firebase-tools": "^11.21.0",
 		"jsonwebtoken": "8.5.1",
 		"micro": "9.3.4",
 		"next": "^13.1.5",
