+++ apps/web/src/pages/forgot-password/EmailInput.tsx
@@ -2,7 +2,7 @@ import BackToLogin from "./BackToLogin"
 import colors from "@Colors"
 import Button from "@Components/shared/Button"
 import TextInput from "@Components/shared/TextInput"
-import { requestForgotPassword } from "@Services/firebase/auth/requestForgotPassword"
+import { requestForgotPassword } from "@kittr/firebase/auth/requestForgotPassword"
 import { useState, Dispatch, SetStateAction, FormEvent } from "react"
 import styled from "styled-components"
 import validator from "validator"
