+++ packages/firebase/auth/getCurrentUser.ts
