+++ packages/firebase/auth/getToken.ts
