+++ packages/firebase/auth/index.ts
