+++ packages/firebase/auth/logOut.ts
