+++ packages/firebase/auth/reauthenticate.ts
