+++ packages/firebase/auth/signUp.ts
