+++ packages/firebase/auth/updateDisplayName.ts
