+++ packages/firebase/auth/updateEmail.ts
