+++ packages/firebase/auth/updatePassword.ts
