+++ packages/firebase/tsconfig.json
@@ -1,0 +1,5 @@
+{
+	"extends": "@kittr/tsconfig/base.json",
+	"include": ["."],
+	"exclude": ["node_modules"]
+}
